requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
Django==5.1.6
django-allauth==65.4.1
django-anymail==12.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==24.3
django-grappelli==4.0.1
django-guardian==2.4.0
django-mysql==4.16.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
iniconfig==2.0.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.3
sslcommerz-lib==1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
